chore: rename files component Input
