+++ src/components/Inputs/TextInput/TextInput.tsx
@@ -2,9 +2,9 @@ import { FC } from 'react'
 
 import { TestingProps } from '@/testing'
 
-import { getClassNames } from './internal/textInput.helpers'
-import s from './internal/textInput.module.scss'
-import { IProps } from './internal/textInput.types'
+import { getClassNames } from './internal/TextInput.helpers'
+import s from './internal/TextInput.module.scss'
+import { IProps } from './internal/TextInput.types'
 
 export const TextInput: FC<IProps & TestingProps> = ({
   value,
