+++ src/components/Inputs/index.tsx
@@ -2,7 +2,7 @@ export { HTTPMethodSelect } from './HTTPMethodSelect'
 export { NumberInput } from './NumberInput'
 export { MillisecondInput } from './MillisecondInput'
 export { SelectInput } from './SelectInput/SelectInput'
-export { TextInput } from './TextInput'
-export { TextAreaInput } from './TextAreaInput'
+export { TextInput } from './TextInput/TextInput'
+export { TextAreaInput } from './TextAreaInput/TextAreaInput'
 export { Toggle } from './Toggle'
 export { URLInput } from './URLInput'
